Added a HUD with buttons
HUD now in the game, shows two buttons, one for producing resource, and
one for spawning minion. Still need to make the buttons functional when
pressed, also more UI elements need to be added to display the spawn
timer and the resource bank for the selected module and add in select
module functionality so that we can spawn different minion types.